move relevant functions to pull.ts
